separate vocabulary and feature extraction code
