+++ returnn/datasets/generating.py
@@ -5,6 +5,7 @@ Some datasets for artificially generated data.
 
 from __future__ import print_function
 
+from returnn.datasets.util.feature_extraction import ExtractAudioFeatures, _get_random_permuted_audio
 from .basic import Dataset, DatasetSeq, convert_data_dims
 from .cached2 import CachedDataset2
 from returnn.util.basic import class_idx_seq_to_1_of_k, CollectionReadCheckCovered, PY3
@@ -12,6 +13,8 @@ from returnn.log import log
 import numpy
 import sys
 import typing
+
+from returnn.datasets.util.vocabulary import *
 
 
 class GeneratingDataset(Dataset):
@@ -992,452 +995,6 @@ class CopyTaskDataset(GeneratingDataset):
 # * nltk.corpus
 
 
-class ExtractAudioFeatures:
-  """
-  Currently uses librosa to extract MFCC/log-mel features.
-  (Alternatives: python_speech_features, talkbox.features.mfcc, librosa)
-  """
-
-  def __init__(self,
-               window_len=0.025, step_len=0.010,
-               num_feature_filters=None, with_delta=False,
-               norm_mean=None, norm_std_dev=None,
-               features="mfcc", feature_options=None, random_permute=None, random_state=None, raw_ogg_opts=None,
-               pre_process=None, post_process=None,
-               sample_rate=None, num_channels=None,
-               peak_normalization=True, preemphasis=None, join_frames=None):
-    """
-    :param float window_len: in seconds
-    :param float step_len: in seconds
-    :param int num_feature_filters:
-    :param bool|int with_delta:
-    :param numpy.ndarray|str|int|float|None norm_mean: if str, will interpret as filename, or "per_seq"
-    :param numpy.ndarray|str|int|float|None norm_std_dev: if str, will interpret as filename, or "per_seq"
-    :param str|function features: "mfcc", "log_mel_filterbank", "log_log_mel_filterbank", "raw", "raw_ogg"
-    :param dict[str]|None feature_options: provide additional parameters for the feature function
-    :param CollectionReadCheckCovered|dict[str]|bool|None random_permute:
-    :param numpy.random.RandomState|None random_state:
-    :param dict[str]|None raw_ogg_opts:
-    :param function|None pre_process:
-    :param function|None post_process:
-    :param int|None sample_rate:
-    :param int|None num_channels: number of channels in audio
-    :param bool peak_normalization: set to False to disable the peak normalization for audio files
-    :param float|None preemphasis: set a preemphasis filter coefficient
-    :param int|None join_frames: concatenate multiple frames together to a superframe
-    :return: float32 data of shape
-    (audio_len // int(step_len * sample_rate), num_channels (optional), (with_delta + 1) * num_feature_filters)
-    :rtype: numpy.ndarray
-    """
-    self.window_len = window_len
-    self.step_len = step_len
-    if num_feature_filters is None:
-      if features == "raw":
-        num_feature_filters = 1
-      elif features == "raw_ogg":
-        raise Exception("you should explicitly specify num_feature_filters (dimension) for raw_ogg")
-      else:
-        num_feature_filters = 40  # was the old default
-    self.num_feature_filters = num_feature_filters
-    self.preemphasis = preemphasis
-    if isinstance(with_delta, bool):
-      with_delta = 1 if with_delta else 0
-    assert isinstance(with_delta, int) and with_delta >= 0
-    self.with_delta = with_delta
-    # join frames needs to be set before norm loading
-    self.join_frames = join_frames
-    if norm_mean is not None:
-      if not isinstance(norm_mean, (int, float)):
-        norm_mean = self._load_feature_vec(norm_mean)
-    if norm_std_dev is not None:
-      if not isinstance(norm_std_dev, (int, float)):
-        norm_std_dev = self._load_feature_vec(norm_std_dev)
-    self.norm_mean = norm_mean
-    self.norm_std_dev = norm_std_dev
-    if random_permute and not isinstance(random_permute, CollectionReadCheckCovered):
-      random_permute = CollectionReadCheckCovered.from_bool_or_dict(random_permute)
-    self.random_permute_opts = random_permute
-    self.random_state = random_state
-    self.features = features
-    self.feature_options = feature_options
-    self.pre_process = pre_process
-    self.post_process = post_process
-    self.sample_rate = sample_rate
-    if num_channels is not None:
-      assert self.features == "raw", "Currently, multiple channels are only supported for raw waveforms"
-      self.num_dim = 3
-    else:
-      self.num_dim = 2
-    self.num_channels = num_channels
-    self.raw_ogg_opts = raw_ogg_opts
-    self.peak_normalization = peak_normalization
-
-  def _load_feature_vec(self, value):
-    """
-    :param str|None value:
-    :return: shape (self.num_inputs,), float32
-    :rtype: numpy.ndarray|str|None
-    """
-    if value is None:
-      return None
-    if isinstance(value, str):
-      if value == "per_seq":
-        return value
-      value = numpy.loadtxt(value)
-    assert isinstance(value, numpy.ndarray)
-    assert value.shape == (self.get_feature_dimension(),)
-    return value.astype("float32")
-
-  def get_audio_features_from_raw_bytes(self, raw_bytes, seq_name=None):
-    """
-    :param io.BytesIO raw_bytes:
-    :param str|None seq_name:
-    :return: shape (time,feature_dim)
-    :rtype: numpy.ndarray
-    """
-    if self.features == "raw_ogg":
-      assert self.with_delta == 0 and self.norm_mean is None and self.norm_std_dev is None
-      # We expect that raw_bytes comes from a Ogg file.
-      try:
-        from returnn.extern.ParseOggVorbis.returnn_import import ParseOggVorbisLib
-      except ImportError:
-        print("Maybe you did not clone the submodule extern/ParseOggVorbis?")
-        raise
-      return ParseOggVorbisLib.get_instance().get_features_from_raw_bytes(
-        raw_bytes=raw_bytes.getvalue(), output_dim=self.num_feature_filters, **(self.raw_ogg_opts or {}))
-
-    # Don't use librosa.load which internally uses audioread which would use Gstreamer as a backend,
-    # which has multiple issues:
-    # https://github.com/beetbox/audioread/issues/62
-    # https://github.com/beetbox/audioread/issues/63
-    # Instead, use PySoundFile, which is also faster. See here for discussions:
-    # https://github.com/beetbox/audioread/issues/64
-    # https://github.com/librosa/librosa/issues/681
-    import soundfile  # noqa  # pip install pysoundfile
-    # integer audio formats are automatically transformed in the range [-1,1]
-    audio, sample_rate = soundfile.read(raw_bytes)
-    return self.get_audio_features(audio=audio, sample_rate=sample_rate, seq_name=seq_name)
-
-  def get_audio_features(self, audio, sample_rate, seq_name=None):
-    """
-    :param numpy.ndarray audio: raw audio samples, shape (audio_len,)
-    :param int sample_rate: e.g. 22050
-    :param str|None seq_name:
-    :return: array (time,dim), dim == self.get_feature_dimension()
-    :rtype: numpy.ndarray
-    """
-    if self.sample_rate is not None:
-      assert sample_rate == self.sample_rate, "currently no conversion implemented..."
-
-    if self.preemphasis:
-      from scipy import signal  # noqa
-      audio = signal.lfilter([1, -self.preemphasis], [1], audio)
-
-    if self.peak_normalization:
-      peak = numpy.max(numpy.abs(audio))
-      if peak != 0.0:
-        audio /= peak
-
-    if self.random_permute_opts and self.random_permute_opts.truth_value:
-      audio = _get_random_permuted_audio(
-        audio=audio,
-        sample_rate=sample_rate,
-        opts=self.random_permute_opts,
-        random_state=self.random_state)
-
-    if self.pre_process:
-      audio = self.pre_process(audio=audio, sample_rate=sample_rate, random_state=self.random_state)
-      assert isinstance(audio, numpy.ndarray) and len(audio.shape) == 1
-
-    if self.features == "raw":
-      assert self.num_feature_filters == 1
-      if audio.ndim == 1:
-        audio = numpy.expand_dims(audio, axis=1)  # add dummy feature axis
-      if self.num_channels is not None:
-        if audio.ndim == 2:
-          audio = numpy.expand_dims(audio, axis=2)  # add dummy feature axis
-        assert audio.shape[1] == self.num_channels
-        assert audio.ndim == 3  # time, channel, feature
-      feature_data = audio.astype("float32")
-
-    else:
-      kwargs = {
-        "sample_rate": sample_rate,
-        "window_len": self.window_len,
-        "step_len": self.step_len,
-        "num_feature_filters": self.num_feature_filters,
-        "audio": audio}
-
-      if self.feature_options is not None:
-        assert isinstance(self.feature_options, dict)
-        kwargs.update(self.feature_options)
-
-      if callable(self.features):
-        feature_data = self.features(random_state=self.random_state, **kwargs)
-      elif self.features == "mfcc":
-        feature_data = _get_audio_features_mfcc(**kwargs)
-      elif self.features == "log_mel_filterbank":
-        feature_data = _get_audio_log_mel_filterbank(**kwargs)
-      elif self.features == "log_log_mel_filterbank":
-        feature_data = _get_audio_log_log_mel_filterbank(**kwargs)
-      elif self.features == "db_mel_filterbank":
-        feature_data = _get_audio_db_mel_filterbank(**kwargs)
-      elif self.features == "linear_spectrogram":
-        feature_data = _get_audio_linear_spectrogram(**kwargs)
-      else:
-        raise Exception("non-supported feature type %r" % (self.features,))
-
-    assert feature_data.ndim == self.num_dim, "got feature data shape %r" % (feature_data.shape,)
-    assert feature_data.shape[-1] == self.num_feature_filters
-
-    if self.with_delta:
-      import librosa  # noqa
-      deltas = [librosa.feature.delta(feature_data, order=i, axis=0).astype("float32")
-                for i in range(1, self.with_delta + 1)]
-      feature_data = numpy.concatenate([feature_data] + deltas, axis=-1)
-      assert feature_data.shape[1] == (self.with_delta + 1) * self.num_feature_filters
-
-    if self.norm_mean is not None:
-      if isinstance(self.norm_mean, str) and self.norm_mean == "per_seq":
-        feature_data -= numpy.mean(feature_data, axis=0, keepdims=True)
-      elif isinstance(self.norm_mean, (int, float)):
-        feature_data -= self.norm_mean
-      else:
-        if self.num_dim == 2:
-          feature_data -= self.norm_mean[numpy.newaxis, :]
-        elif self.num_dim == 3:
-          feature_data -= self.norm_mean[numpy.newaxis, numpy.newaxis, :]
-        else:
-          assert False, "Unexpected number of dimensions: {}".format(self.num_dim)
-
-    if self.norm_std_dev is not None:
-      if isinstance(self.norm_std_dev, str) and self.norm_std_dev == "per_seq":
-        feature_data /= numpy.maximum(numpy.std(feature_data, axis=0, keepdims=True), 1e-2)
-      elif isinstance(self.norm_std_dev, (int, float)):
-        feature_data /= self.norm_std_dev
-      else:
-        if self.num_dim == 2:
-          feature_data /= self.norm_std_dev[numpy.newaxis, :]
-        elif self.num_dim == 3:
-          feature_data /= self.norm_std_dev[numpy.newaxis, numpy.newaxis, :]
-        else:
-          assert False, "Unexpected number of dimensions: {}".format(self.num_dim)
-
-    if self.join_frames is not None:
-      pad_len = self.join_frames - (feature_data.shape[0] % self.join_frames)
-      pad_len = pad_len % self.join_frames
-      new_len = feature_data.shape[0] + pad_len
-      if self.num_channels is None:
-        new_shape = (new_len // self.join_frames, feature_data.shape[-1] * self.join_frames)
-        pad_width = ((0, pad_len), (0, 0))
-      else:
-        new_shape = (new_len // self.join_frames, self.num_channels, feature_data.shape[-1] * self.join_frames)
-        pad_width = ((0, pad_len), (0, 0), (0, 0))
-      feature_data = numpy.pad(feature_data, pad_width=pad_width, mode="edge")
-      feature_data = numpy.reshape(feature_data, newshape=new_shape, order='C')
-
-    assert feature_data.shape[-1] == self.get_feature_dimension()
-    if self.post_process:
-      feature_data = self.post_process(feature_data, seq_name=seq_name)
-      assert isinstance(feature_data, numpy.ndarray) and feature_data.ndim == self.num_dim
-      assert feature_data.shape[-1] == self.get_feature_dimension()
-    return feature_data
-
-  def get_feature_dimension(self):
-    """
-    :rtype: int
-    """
-    return (self.with_delta + 1) * self.num_feature_filters * (self.join_frames or 1)
-
-
-def _get_audio_linear_spectrogram(audio, sample_rate, window_len=0.025, step_len=0.010, num_feature_filters=512):
-  """
-  Computes linear spectrogram features from an audio signal.
-  Drops the DC component.
-
-  :param numpy.ndarray audio: raw audio samples, shape (audio_len,)
-  :param int sample_rate: e.g. 22050
-  :param float window_len: in seconds
-  :param float step_len: in seconds
-  :return: (audio_len // int(step_len * sample_rate), num_feature_filters), float32
-  :rtype: numpy.ndarray
-  """
-  import librosa  # noqa
-
-  min_n_fft = int(window_len * sample_rate)
-  assert num_feature_filters*2 >= min_n_fft
-  assert num_feature_filters % 2 == 0
-
-  spectrogram = numpy.abs(librosa.core.stft(
-    audio, hop_length=int(step_len * sample_rate),
-    win_length=int(window_len * sample_rate), n_fft=num_feature_filters*2))
-
-  # remove the DC part
-  spectrogram = spectrogram[1:]
-
-  assert spectrogram.shape[0] == num_feature_filters
-  spectrogram = spectrogram.transpose().astype("float32")  # (time, dim)
-  return spectrogram
-
-
-def _get_audio_features_mfcc(audio, sample_rate, window_len=0.025, step_len=0.010, num_feature_filters=40):
-  """
-  :param numpy.ndarray audio: raw audio samples, shape (audio_len,)
-  :param int sample_rate: e.g. 22050
-  :param float window_len: in seconds
-  :param float step_len: in seconds
-  :param int num_feature_filters:
-  :return: (audio_len // int(step_len * sample_rate), num_feature_filters), float32
-  :rtype: numpy.ndarray
-  """
-  import librosa  # noqa
-  features = librosa.feature.mfcc(
-    audio, sr=sample_rate,
-    n_mfcc=num_feature_filters,
-    hop_length=int(step_len * sample_rate), n_fft=int(window_len * sample_rate))
-  librosa_version = librosa.__version__.split(".")
-  if int(librosa_version[0]) >= 1 or (int(librosa_version[0]) == 0 and int(librosa_version[1]) >= 7):
-    rms_func = librosa.feature.rms
-  else:
-    rms_func = librosa.feature.rmse  # noqa
-  energy = rms_func(
-    audio,
-    hop_length=int(step_len * sample_rate), frame_length=int(window_len * sample_rate))
-  features[0] = energy  # replace first MFCC with energy, per convention
-  assert features.shape[0] == num_feature_filters  # (dim, time)
-  features = features.transpose().astype("float32")  # (time, dim)
-  return features
-
-
-def _get_audio_log_mel_filterbank(audio, sample_rate, window_len=0.025, step_len=0.010, num_feature_filters=80):
-  """
-  Computes log Mel-filterbank features from an audio signal.
-  References:
-
-    https://github.com/jameslyons/python_speech_features/blob/master/python_speech_features/base.py
-    https://github.com/tensorflow/tensor2tensor/blob/master/tensor2tensor/data_generators/speech_recognition.py
-
-  :param numpy.ndarray audio: raw audio samples, shape (audio_len,)
-  :param int sample_rate: e.g. 22050
-  :param float window_len: in seconds
-  :param float step_len: in seconds
-  :param int num_feature_filters:
-  :return: (audio_len // int(step_len * sample_rate), num_feature_filters), float32
-  :rtype: numpy.ndarray
-  """
-  import librosa  # noqa
-  mel_filterbank = librosa.feature.melspectrogram(
-    audio, sr=sample_rate,
-    n_mels=num_feature_filters,
-    hop_length=int(step_len * sample_rate), n_fft=int(window_len * sample_rate))
-  log_noise_floor = 1e-3  # prevent numeric overflow in log
-  log_mel_filterbank = numpy.log(numpy.maximum(log_noise_floor, mel_filterbank))
-  assert log_mel_filterbank.shape[0] == num_feature_filters
-  log_mel_filterbank = log_mel_filterbank.transpose().astype("float32")  # (time, dim)
-  return log_mel_filterbank
-
-
-def _get_audio_db_mel_filterbank(audio, sample_rate,
-                                 window_len=0.025, step_len=0.010, num_feature_filters=80,
-                                 fmin=0, fmax=None, min_amp=1e-10):
-  """
-  Computes log Mel-filterbank features in dezibel values from an audio signal.
-  Provides adjustable minimum frequency and minimual amplitude clipping
-
-  :param numpy.ndarray audio: raw audio samples, shape (audio_len,)
-  :param int sample_rate: e.g. 22050
-  :param float window_len: in seconds
-  :param float step_len: in seconds
-  :param int num_feature_filters: number of mel-filterbanks
-  :param int fmin: minimum frequency covered by mel filters
-  :param int|None fmax: maximum frequency covered by mel filters
-  :param int min_amp: silence clipping for small amplitudes
-  :return: (audio_len // int(step_len * sample_rate), num_feature_filters), float32
-  :rtype: numpy.ndarray
-  """
-  # noinspection PyPackageRequirements
-  assert fmin >= 0
-  assert min_amp > 0
-
-  import librosa  # noqa
-  mel_filterbank = librosa.feature.melspectrogram(
-    audio, sr=sample_rate,
-    n_mels=num_feature_filters,
-    hop_length=int(step_len * sample_rate),
-    n_fft=int(window_len * sample_rate),
-    fmin=fmin, fmax=fmax,
-   )
-
-  log_mel_filterbank = 20 * numpy.log10(numpy.maximum(min_amp, mel_filterbank))
-  assert log_mel_filterbank.shape[0] == num_feature_filters
-  log_mel_filterbank = log_mel_filterbank.transpose().astype("float32")  # (time, dim)
-  return log_mel_filterbank
-
-
-def _get_audio_log_log_mel_filterbank(audio, sample_rate, window_len=0.025, step_len=0.010, num_feature_filters=80):
-  """
-  Computes log-log Mel-filterbank features from an audio signal.
-  References:
-
-    https://github.com/jameslyons/python_speech_features/blob/master/python_speech_features/base.py
-    https://github.com/tensorflow/tensor2tensor/blob/master/tensor2tensor/data_generators/speech_recognition.py
-
-  :param numpy.ndarray audio: raw audio samples, shape (audio_len,)
-  :param int sample_rate: e.g. 22050
-  :param float window_len: in seconds
-  :param float step_len: in seconds
-  :param int num_feature_filters:
-  :return: (audio_len // int(step_len * sample_rate), num_feature_filters), float32
-  :rtype: numpy.ndarray
-  """
-  import librosa  # noqa
-  mel_filterbank = librosa.feature.melspectrogram(
-    audio, sr=sample_rate,
-    n_mels=num_feature_filters,
-    hop_length=int(step_len * sample_rate), n_fft=int(window_len * sample_rate))
-  log_noise_floor = 1e-3  # prevent numeric overflow in log
-  log_mel_filterbank = numpy.log(numpy.maximum(log_noise_floor, mel_filterbank))
-  log_log_mel_filterbank = librosa.core.amplitude_to_db(log_mel_filterbank)
-  assert log_log_mel_filterbank.shape[0] == num_feature_filters
-  log_log_mel_filterbank = log_log_mel_filterbank.transpose().astype("float32")  # (time, dim)
-  return log_log_mel_filterbank
-
-
-def _get_random_permuted_audio(audio, sample_rate, opts, random_state):
-  """
-  :param numpy.ndarray audio: raw time signal
-  :param int sample_rate:
-  :param CollectionReadCheckCovered opts:
-  :param numpy.random.RandomState random_state:
-  :return: audio randomly permuted
-  :rtype: numpy.ndarray
-  """
-  import librosa  # noqa
-  import scipy.ndimage  # noqa
-  import warnings
-  audio = audio * random_state.uniform(opts.get("rnd_scale_lower", 0.8), opts.get("rnd_scale_upper", 1.0))
-  if opts.get("rnd_zoom_switch", 1.) > 0.:
-    opts.get("rnd_zoom_lower"), opts.get("rnd_zoom_upper"), opts.get("rnd_zoom_order")  # Mark as read.
-  if random_state.uniform(0.0, 1.0) < opts.get("rnd_zoom_switch", 0.2):
-    with warnings.catch_warnings():
-      warnings.simplefilter("ignore")
-      # Alternative: scipy.interpolate.interp2d
-      factor = random_state.uniform(opts.get("rnd_zoom_lower", 0.9), opts.get("rnd_zoom_upper", 1.1))
-      audio = scipy.ndimage.zoom(audio, factor, order=opts.get("rnd_zoom_order", 3))
-  if opts.get("rnd_stretch_switch", 1.) > 0.:
-    opts.get("rnd_stretch_lower"), opts.get("rnd_stretch_upper")  # Mark as read.
-  if random_state.uniform(0.0, 1.0) < opts.get("rnd_stretch_switch", 0.2):
-    rate = random_state.uniform(opts.get("rnd_stretch_lower", 0.9), opts.get("rnd_stretch_upper", 1.2))
-    audio = librosa.effects.time_stretch(audio, rate=rate)
-  if opts.get("rnd_pitch_switch", 1.) > 0.:
-    opts.get("rnd_pitch_lower"), opts.get("rnd_pitch_upper", 1.)  # Mark as read.
-  if random_state.uniform(0.0, 1.0) < opts.get("rnd_pitch_switch", 0.2):
-    n_steps = random_state.uniform(opts.get("rnd_pitch_lower", -1.), opts.get("rnd_pitch_upper", 1.))
-    audio = librosa.effects.pitch_shift(audio, sr=sample_rate, n_steps=n_steps)
-  opts.assert_all_read()
-  return audio
-
-
 class TimitDataset(CachedDataset2):
   """
   DARPA TIMIT Acoustic-Phonetic Continuous Speech Corpus.
@@ -1875,351 +1432,6 @@ class NltkTimitDataset(TimitDataset):
 
   def _read_phone_seq(self, seq_tag):
     return self._data_reader.phones(seq_tag)
-
-
-class Vocabulary(object):
-  """
-  Represents a vocabulary (set of words, and their ids).
-  Used by :class:`BytePairEncoding`.
-  """
-
-  _cache = {}  # filename -> vocab dict, labels dict (see _parse_vocab)
-
-  @classmethod
-  def create_vocab(cls, **opts):
-    """
-    :param opts: kwargs for class
-    :rtype: Vocabulary|BytePairEncoding|CharacterTargets
-    """
-    opts = opts.copy()
-    clz = cls
-    if "class" in opts:
-      class_name = opts.pop("class")
-      clz = globals()[class_name]
-      assert issubclass(clz, Vocabulary), "class %r %r is not a subclass of %r" % (class_name, clz, cls)
-    elif "bpe_file" in opts:
-      clz = BytePairEncoding
-    return clz(**opts)
-
-  def __init__(self, vocab_file, seq_postfix=None, unknown_label="UNK", num_labels=None):
-    """
-    :param str|None vocab_file:
-    :param str|None unknown_label:
-    :param int num_labels: just for verification
-    :param list[int]|None seq_postfix: labels will be added to the seq in self.get_seq
-    """
-    self.vocab_file = vocab_file
-    self.unknown_label = unknown_label
-    self.num_labels = None  # type: typing.Optional[int]  # will be set by _parse_vocab
-    self.vocab = None  # type: typing.Optional[typing.Dict[str,int]]  # label->idx
-    self.labels = None  # type: typing.Optional[typing.List[str]]
-    self._parse_vocab()
-    if num_labels is not None:
-      assert self.num_labels == num_labels
-    self.unknown_label_id = self.vocab[self.unknown_label] if self.unknown_label is not None else None
-    self.seq_postfix = seq_postfix or []
-
-  def __repr__(self):
-    return "Vocabulary(%r, num_labels=%s, unknown_label=%r)" % (self.vocab_file, self.num_labels, self.unknown_label)
-
-  def set_random_seed(self, seed):
-    """
-    This can be called for a new epoch or so.
-    Usually it has no effect, as there is no randomness.
-    However, some vocab class could introduce some sampling process.
-
-    :param int seed:
-    """
-    pass  # usually there is no randomness, so ignore
-
-  def _parse_vocab(self):
-    """
-    Sets self.vocab, self.labels, self.num_labels.
-    """
-    filename = self.vocab_file
-    import pickle
-    if filename in self._cache:
-      self.vocab, self.labels = self._cache[filename]
-      assert self.unknown_label is None or self.unknown_label in self.vocab
-      self.num_labels = len(self.labels)
-    else:
-      if filename[-4:] == ".pkl":
-        d = pickle.load(open(filename, "rb"))
-      else:
-        d = eval(open(filename, "r").read())
-        if not PY3:
-          # Any utf8 string will not be a unicode string automatically, so enforce this.
-          assert isinstance(d, dict)
-          from returnn.util.basic import py2_utf8_str_to_unicode
-          d = {py2_utf8_str_to_unicode(s): i for (s, i) in d.items()}
-      assert isinstance(d, dict)
-      assert self.unknown_label is None or self.unknown_label in d
-      labels = {idx: label for (label, idx) in sorted(d.items())}
-      min_label, max_label, num_labels = min(labels), max(labels), len(labels)
-      assert 0 == min_label
-      if num_labels - 1 < max_label:
-        print("Vocab error: not all indices used? max label: %i" % max_label, file=log.v1)
-        print("unused labels: %r" % ([i for i in range(max_label + 1) if i not in labels],), file=log.v2)
-      assert num_labels - 1 == max_label
-      self.num_labels = len(labels)
-      self.vocab = d
-      self.labels = [label for (idx, label) in sorted(labels.items())]
-      self._cache[filename] = (self.vocab, self.labels)
-
-  @classmethod
-  def create_vocab_dict_from_labels(cls, labels):
-    """
-    This is exactly the format which we expect when we read it in self._parse_vocab.
-
-    :param list[str] labels:
-    :rtype: dict[str,int]
-    """
-    d = {label: idx for (idx, label) in enumerate(labels)}
-    assert len(d) == len(labels), "some labels are provided multiple times"
-    return d
-
-  def tf_get_init_variable_func(self, var):
-    """
-    :param tensorflow.Variable var:
-    :rtype: (tensorflow.Session)->None
-    """
-    import tensorflow as tf
-    from returnn.tf.util.basic import VariableAssigner
-    assert isinstance(var, tf.Variable)
-    assert var.dtype.base_dtype == tf.string
-    assert var.shape.as_list() == [self.num_labels]
-    assert len(self.labels) == self.num_labels
-
-    def init_vocab_var(session):
-      """
-      :param tensorflow.Session session:
-      """
-      VariableAssigner(var).assign(session=session, value=self.labels)
-
-    return init_vocab_var
-
-  def get_seq(self, sentence):
-    """
-    :param str sentence: assumed to be seq of vocab entries separated by whitespace
-    :rtype: list[int]
-    """
-    segments = sentence.split()
-    return self.get_seq_indices(segments) + self.seq_postfix
-
-  def get_seq_indices(self, seq):
-    """
-    :param list[str] seq:
-    :rtype: list[int]
-    """
-    if self.unknown_label is not None:
-      return [self.vocab.get(k, self.unknown_label_id) for k in seq]
-    return [self.vocab[k] for k in seq]
-
-  def get_seq_labels(self, seq):
-    """
-    :param list[int] seq:
-    :rtype: str
-    """
-    return " ".join(map(self.labels.__getitem__, seq))
-
-
-class BytePairEncoding(Vocabulary):
-  """
-  Vocab based on Byte-Pair-Encoding (BPE).
-  This will encode the text on-the-fly with BPE.
-
-  Reference:
-  Rico Sennrich, Barry Haddow and Alexandra Birch (2016). Neural Machine Translation of Rare Words with Subword Units.
-  Proceedings of the 54th Annual Meeting of the Association for Computational Linguistics (ACL 2016). Berlin, Germany.
-  """
-
-  def __init__(self, vocab_file, bpe_file, seq_postfix=None, unknown_label="UNK"):
-    """
-    :param str vocab_file:
-    :param str bpe_file:
-    :param list[int]|None seq_postfix: labels will be added to the seq in self.get_seq
-    :param str|None unknown_label:
-    """
-    super(BytePairEncoding, self).__init__(vocab_file=vocab_file, seq_postfix=seq_postfix, unknown_label=unknown_label)
-    from returnn.util.bpe import StandardBytePairEncoder
-    self.bpe = StandardBytePairEncoder(bpe_codes_file=bpe_file, labels=self.labels)
-
-  def get_seq(self, sentence):
-    """
-    :param str sentence:
-    :rtype: list[int]
-    """
-    segments = self.bpe.segment_sentence(sentence)
-    seq = self.get_seq_indices(segments)
-    return seq + self.seq_postfix
-
-
-class SamplingBytePairEncoding(Vocabulary):
-  """
-  Vocab based on Byte-Pair-Encoding (BPE).
-  Like :class:`BytePairEncoding`, but here we randomly sample from different possible BPE splits.
-  This will encode the text on-the-fly with BPE.
-  """
-
-  def __init__(self, vocab_file, breadth_prob, seq_postfix=None, unknown_label="UNK"):
-    """
-    :param str vocab_file:
-    :param float breadth_prob:
-    :param list[int]|None seq_postfix: labels will be added to the seq in self.get_seq
-    :param str|None unknown_label:
-    """
-    super(SamplingBytePairEncoding, self).__init__(
-      vocab_file=vocab_file, seq_postfix=seq_postfix, unknown_label=unknown_label)
-    from returnn.util.bpe import SamplingBytePairEncoder
-    self.rnd = numpy.random.RandomState(0)
-    self.bpe = SamplingBytePairEncoder(
-      labels=self.labels, breadth_prob=breadth_prob, rnd=self.rnd, unknown_label=unknown_label)
-
-  def set_random_seed(self, seed):
-    """
-    :param int seed:
-    """
-    self.rnd.seed(seed)
-
-  def get_seq(self, sentence):
-    """
-    :param str sentence:
-    :rtype: list[int]
-    """
-    segments = self.bpe.segment_sentence(sentence)
-    seq = self.get_seq_indices(segments)
-    return seq + self.seq_postfix
-
-
-class SentencePieces(Vocabulary):
-  """
-  Uses the SentencePiece software,
-  which supports different kind of subword units (including BPE, unigram, ...).
-
-  https://github.com/google/sentencepiece/
-  https://github.com/google/sentencepiece/tree/master/python
-
-  Dependency::
-
-    pip3 install --user sentencepiece
-
-  """
-
-  def __init__(self, **opts):
-    """
-    :param str model_file: The sentencepiece model file path.
-    :param str model_proto: The sentencepiece model serialized proto.
-    :param type out_type: output type. int or str. (Default = int)
-    :param bool add_bos: Add <s> to the result (Default = false)
-    :param bool add_eos: Add </s> to the result (Default = false)
-      <s>/</s> is added after reversing (if enabled).
-    :param bool reverse: Reverses the tokenized sequence (Default = false)
-    :param bool enable_sampling: (Default = false)
-    :param int nbest_size: sampling parameters for unigram. Invalid for BPE-Dropout.
-      nbest_size = {0,1}: No sampling is performed.
-      nbest_size > 1: samples from the nbest_size results.
-      nbest_size < 0: (Default). assuming that nbest_size is infinite and samples
-        from the all hypothesis (lattice) using
-        forward-filtering-and-backward-sampling algorithm.
-    :param float alpha: Soothing parameter for unigram sampling, and dropout probability of
-      merge operations for BPE-dropout. (Default = 0.1)
-    """
-    import sentencepiece as spm  # noqa
-    self._opts = opts
-    self._cache_key = opts.get("model_file", None)
-    self.sp = spm.SentencePieceProcessor(**opts)  # noqa
-    super(SentencePieces, self).__init__(
-      vocab_file=None, seq_postfix=None, unknown_label=self.sp.IdToPiece(self.sp.unk_id()))
-
-  def __repr__(self):
-    return "SentencePieces(%r)" % (self._opts,)
-
-  def _parse_vocab(self):
-    self.num_labels = self.sp.vocab_size()
-    if self._cache_key and self._cache_key in self._cache:
-      self.vocab, self.labels = self._cache[self._cache_key]
-      assert self.unknown_label in self.vocab and self.num_labels == len(self.vocab) == len(self.labels)
-      return
-    self.labels = [self.sp.id_to_piece(i) for i in range(self.num_labels)]  # noqa
-    self.vocab = {label: i for (i, label) in enumerate(self.labels)}
-    if self._cache_key:
-      self._cache[self._cache_key] = (self.vocab, self.labels)
-
-  def set_random_seed(self, seed):
-    """
-    :param int seed:
-    """
-    # Unfortunately, there is only a global seed,
-    # and also, it will only be used for new threads
-    # where the random generator was not used yet...
-    # https://github.com/google/sentencepiece/issues/635
-    import sentencepiece as spm  # noqa
-    spm.set_random_generator_seed(seed)
-
-  def get_seq(self, sentence):
-    """
-    :param str sentence: assumed to be seq of vocab entries separated by whitespace
-    :rtype: list[int]
-    """
-    return self.sp.encode(sentence, out_type=int)  # noqa
-
-
-class CharacterTargets(Vocabulary):
-  """
-  Uses characters as target labels.
-  Also see :class:`Utf8ByteTargets`.
-  """
-
-  def __init__(self, vocab_file, seq_postfix=None, unknown_label="@"):
-    """
-    :param str vocab_file:
-    :param list[int]|None seq_postfix: labels will be added to the seq in self.get_seq
-    :param str|None unknown_label:
-    """
-    super(CharacterTargets, self).__init__(vocab_file=vocab_file, seq_postfix=seq_postfix, unknown_label=unknown_label)
-
-  def get_seq(self, sentence):
-    """
-    :param str sentence:
-    :rtype: list[int]
-    """
-    if self.unknown_label is not None:
-      seq = [self.vocab.get(k, self.unknown_label_id) for k in sentence]
-    else:
-      seq = [self.vocab[k] for k in sentence]
-    return seq + self.seq_postfix
-
-
-class Utf8ByteTargets(Vocabulary):
-  """
-  Uses bytes as target labels from UTF8 encoded text. All bytes (0-255) are allowed.
-  Also see :class:`CharacterTargets`.
-  """
-
-  def __init__(self, seq_postfix=None):
-    """
-    :param list[int]|None seq_postfix: labels will be added to the seq in self.get_seq
-    """
-    super(Utf8ByteTargets, self).__init__(vocab_file=None, seq_postfix=seq_postfix, unknown_label=None)
-
-  def _parse_vocab(self):
-    """
-    Sets self.vocab, self.labels, self.num_labels.
-    """
-    self.vocab = {chr(i): i for i in range(256)}
-    self.labels = [chr(i) for i in range(256)]
-    self.num_labels = 256
-
-  def get_seq(self, sentence):
-    """
-    :param str sentence:
-    :rtype: list[int]
-    """
-    if sys.version_info[0] >= 3:
-      seq = list(sentence.encode("utf8"))
-    else:
-      seq = list(bytearray(sentence.encode("utf8")))
-    return seq + self.seq_postfix
 
 
 class BlissDataset(CachedDataset2):
